// apps/frontend/src/app/api/generated/model/worksiteResponse.ts
/**
 * Janus API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WorksiteScope } from './worksiteScope';


export interface WorksiteResponse { 
    code: string;
    name: string;
    timeZone: string;
    scope: WorksiteScope;
    description?: string | null;
    address?: string | null;
    ownerEmployeeEmail?: string | null;
    active: boolean;
}